タスクの実装部分のファイルを整理
